+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+Added
+  * AZ-1000: Add an option to enable or disable UAI Private DNS Zone role assignment
+  * AZ-1001: Add default `no_proxy_url_list` and some information about the parameter
+
 # v7.4.0 - 2023-02-10
 
 Changed